refactor: make deserializers internal
