+++ src/data_factory_testing_framework/repositories/data_factory_repository_factory.py
@@ -1,6 +1,6 @@
 import os
 
-from data_factory_testing_framework.deserializers._deserializer_data_factory import (
+from data_factory_testing_framework._deserializers._deserializer_data_factory import (
     parse_data_factory_pipeline_from_pipeline_json,
 )
 from data_factory_testing_framework.models.pipeline import Pipeline
